==> Notebook/1-Base de données piezos.ipynb <==
import requests
import pandas as pd
import numpy as np
import sqlite3
from sqlalchemy import create_engine 
import os

# Récupération des données

## Requête API et création du dataframe

def f_list_piezzo(departement:list, nbre_mesures:int=50) -> list:
    """
    departement : liste des départements recherchés
    return : liste des code_bss et autres informations
    """
    requete_dpt = ""
    for dpt in departement:
        if dpt < 10 :
            requete_dpt += '0' + str(dpt) + '%2C'
        else :
            requete_dpt += str(dpt) + '%2C'
    requete_dpt = requete_dpt[0:-3]    
        
    url = f"https://hubeau.eaufrance.fr/api/v1/niveaux_nappes/stations?code_departement={requete_dpt}&format=json&nb_mesures_piezo_min={nbre_mesures}&size=10000"
    print(url)
    get= requests.get(f"{url}")
    data_piezo = pd.DataFrame(columns=['code_bss',
                                       'date_debut',
                                      'date_fin',
                                       'nb_mesures_piezo',
                                       'code_dpt',
                                      'code_commune_insee',
                                      'nom_commune',
                                      'latitude',
                                      'longitude',
                                      'altitude_station',
                                      'profondeur_investigation',
                                      "codes_bdlisa",
                                      "libelle_pe",
                                      "codes_masse_eau_edl",
                                      "noms_masse_eau_edl"])
    data_piezo["code_bss"] = [x['code_bss'].replace("/","_") for x in get.json()['data']]
    data_piezo["date_debut"] = [pd.to_datetime(x['date_debut_mesure'], format="%Y/%m/%d") for x in get.json()['data']]
    data_piezo["date_fin"] = [pd.to_datetime(x['date_fin_mesure'], format="%Y/%m/%d") for x in get.json()['data']]
    data_piezo["nb_mesures_piezo"] = [x['nb_mesures_piezo'] for x in get.json()['data']]    
    data_piezo["code_dpt"] = [x['code_departement'] for x in get.json()['data']]    
    data_piezo["code_commune_insee"] = [x['code_commune_insee'] for x in get.json()['data']]
    data_piezo["nom_commune"] = [x['nom_commune'] for x in get.json()['data']]
    data_piezo["latitude"] = [x['y'] for x in get.json()['data']]
    data_piezo["longitude"] = [x['x'] for x in get.json()['data']]
    data_piezo["altitude_station"] = [float(x['altitude_station']) for x in get.json()['data']]
    data_piezo["profondeur_investigation"] = [x['profondeur_investigation'] for x in get.json()['data']] 
    
    # Gestion du format tableau et des Nan pour les colonnes suivantes :
    data_piezo["codes_bdlisa"] = [x['codes_bdlisa'] for x in get.json()['data']]
    bdlisa = []
    for item in list(data_piezo["codes_bdlisa"]):
        if item == None:
            bdlisa.append(None)
        else :
            bdlisa.append(item[0])
    data_piezo["codes_bdlisa"] = np.array(bdlisa)
    
    data_piezo["libelle_pe"] = [x['libelle_pe'] for x in get.json()['data']]
    libelle_pe = []
    for item in list(data_piezo["libelle_pe"]):
        if item == None:
            libelle_pe.append(None)
        else :
            libelle_pe.append(item[0])
    data_piezo["libelle_pe"] = np.array(libelle_pe)
    
    data_piezo["codes_masse_eau_edl"] = [x['codes_masse_eau_edl'] for x in get.json()['data']]
    codes_masse_eau_edl = []
    for item in list(data_piezo["codes_masse_eau_edl"]):
        if item == None:
            codes_masse_eau_edl.append(None)
        else :
            codes_masse_eau_edl.append(item[0])
    data_piezo["codes_masse_eau_edl"] = np.array(codes_masse_eau_edl)
    
    
    data_piezo["noms_masse_eau_edl"] = [x['noms_masse_eau_edl'] for x in get.json()['data']]
    noms_masse_eau_edl = []
    for item in list(data_piezo["noms_masse_eau_edl"]):
        if item == None:
            noms_masse_eau_edl.append(None)
        else :
            noms_masse_eau_edl.append(item[0])
    data_piezo["noms_masse_eau_edl"] = np.array(noms_masse_eau_edl)
    
    return data_piezo


# Création d'une liste de départements de 1 à 95.
liste = [_ for _ in range(1,96)]

# Requete API et création du dataframe Pandas
df_piezo = f_list_piezzo(liste, nbre_mesures=50).set_index('code_bss')
df_piezo.head()

df_piezo.shape

## Recherche d'un piezo en particulier

print(df_piezo.loc[df_piezo.index=="04994X0528_FPZ","noms_masse_eau_edl"])

# Création de la base de données

## Création directe à partir du dataframe avec sqlalchemy

# Affichage des colonnes du dataframe créé
df_piezo.columns

# Création de la base
engine = create_engine('sqlite:///../data/liste_piezos.db',  
                   echo = False, encoding='utf8')

if not os.path.isfile("../data/liste_piezos.db"):
    try : 
        engine.execute("""
                        CREATE TABLE IF NOT EXISTS data_piezo (
                        code_bss TEXT PRIMARY KEY ON CONFLICT IGNORE,
                        date_debut DATE,
                        date_fin DATE,
                        nb_mesures_piezo MEDIUMINT,
                        code_dpt VARCHAR(3),
                        code_commune_insee VARCHAR (6),
                        nom_commune VARCHAR(50),
                        latitude FLOAT,
                        longitude FLOAT,
                        altitude_station FLOAT,
                        profondeur_investigation FLOAT,
                        codes_bdlisa VARCHAR(50),
                        libelle_pe VARCHAR (250),
                        codes_masse_eau_edl VARCHAR (20),
                        noms_masse_eau_edl TEXT,           
                        fichier_piezo TINYINT DEFAULT 0,
                        fichier_era5 TINYINT DEFAULT 0,
                        fichier_combined TINYINT DEFAULT 0,
                        commentaire TEXT
                        )
                        """)
        print("La base de données liste_piezos a été créée.")
    except InterfaceError as e :
        print("Problème durant la création : ", e)
else :
    print("La base de données liste_piezos existe déjà.")

# Ajout des données dans la base de données
if not os.path.isfile("../data/liste_piezos.db"):
    df_piezo.to_sql('data_piezo', 
                    con = engine,
                    if_exists='append')

    nb_lignes = engine.execute("SELECT COUNT (*) FROM data_piezo").fetchall()[0][0]
    print(f"La base de données contient {nb_lignes} lignes.")
    engine.dispose()

else :
    reponse = str.lower(input("""Voulez vous écraser la base éxistante ? (O/N)"""))
    if reponse == 'o':
        df_piezo.to_sql('data_piezo', 
                    con = engine,
                    if_exists='append')

        nb_lignes = engine.execute("SELECT COUNT (*) FROM data_piezo").fetchall()[0][0]
        print(f"La base de données contient {nb_lignes} lignes.")
        engine.dispose()
    else:
        print("La base n'a pas été modifié.")

## Fichiers disponibles pour les piezos ?

def requete_sql (requete) :
    try:
        connexion = sqlite3.connect('./data/liste_piezos.db')
        curseur = connexion.cursor()
        print("Connexion réussie à SQLite")
        curseur.execute(requete)
        connexion.commit()
        resultat = curseur.fetchall()
        curseur.close()
        connexion.close()
        print("Connexion SQLite est fermée")
        return resultat
    except sqlite3.Error as error:
        print("Erreur lors du mis à jour dans la table", error)

bss_piezo = pd.Series([_.replace(".csv","") for _ in os.listdir(r"./data/piezo") if _.endswith(r".csv")], name="piezo")    
bss_era5 = pd.Series([_.replace(".pkl","") for _ in os.listdir(r"./data/era5") if _.endswith(r".pkl")], name="era5")
bss_combined = np.array([_[5:-4] for _ in os.listdir(r"./data/combined_data/") if _.endswith(r".csv")])
    
requete_fichier_piezo = f"""
        UPDATE data_piezo
        SET fichier_piezo = 1
        WHERE code_bss IN {tuple(bss_piezo)}
         """
requete_sql(requete_fichier_piezo)

requete_fichier_era5 = f"""
        UPDATE data_piezo
        SET fichier_era5 = 1
        WHERE code_bss IN {tuple(bss_era5)}
         """
requete_sql(requete_fichier_era5)

requete_fichier_combined = f"""
        UPDATE data_piezo
        SET fichier_combined = 1
        WHERE code_bss IN {tuple(bss_combined)}
         """
requete_sql(requete_fichier_combined)

requete = f"""
        SELECT COUNT(*)
        FROM data_piezo
        WHERE fichier_combined IS 1
        """

requete_sql(requete)

# Table Départements et régions

df = pd.read_csv("./data/DPT_REG.csv", sep=";", index_col="departement")

df.head()

connexion = sqlite3.connect('./data/liste_piezos.db')
df.to_sql("code_region", connexion, if_exists='replace', index=True)
