==> Notebook/2-Import Hub_eau.ipynb <==
import sqlite3
import pandas as pd
import numpy as np
import os
import requests

def requete_sql (requete) :
    try:
        connexion = sqlite3.connect('../data/liste_piezos.db')
        curseur = connexion.cursor()
        print("Connexion réussie à SQLite")
        curseur.execute(requete)
        connexion.commit()
        resultat = curseur.fetchall()
        curseur.close()
        connexion.close()
        print("Connexion SQLite est fermée")
        return resultat
    except sqlite3.Error as error:
        print("Erreur lors du mis à jour dans la table", error)

requete = "select code_bss from data_piezo"
liste_bss = [piezo[0] for piezo in requete_sql (requete)]

# Connexion à la bdd et création d'un dataframe

# Récupération des données 

connexion = sqlite3.connect('../data/liste_piezos.db')
curseur = connexion.cursor()
requete = """
    SELECT * FROM data_piezo    
    WHERE date_debut < '2010-07-01'
    AND date_fin > '2010-07-01'
"""
df_piezo = pd.read_sql_query(requete, connexion, index_col='code_bss')
curseur.close()
connexion.close()

# Conversion des dates
df_piezo["date_debut"] = pd.to_datetime(df_piezo["date_debut"])
df_piezo["date_fin"] = pd.to_datetime(df_piezo["date_fin"])

df_piezo.head()

df_piezo.shape

# Fonction de collecte de la chronique d'un piezo

def f_data_piezo (code_bss:str, date_index:bool=True):
    """
    code_bss : code bss brut
    date_index : (default: True), date en index du dataframe
    return : csv file
    """
    # Nom de fichier
    file  = f'{code_bss}.csv'
    if not os.path.isfile(f'../data/piezo/{file}') : #si le fichier n'existe pas
        code_bss_transfo = code_bss.replace("_", "%2F", 1) #Nécessaire pour construire l'URL
        url = f"http://hubeau.eaufrance.fr/api/v1/niveaux_nappes/chroniques?code_bss={code_bss_transfo}&size=20000"
        print(url)
        try :
            get = requests.get(url)
            json = get.json()
            data = json['data']
            #Construction du dataframe
            df = pd.DataFrame(columns=['date_mesure', 'piezo'])
            df['date_mesure'] = pd.concat([pd.DataFrame([mesure['date_mesure']], columns=['date_mesure']) for mesure in data], ignore_index=True)
            df["piezo"] = pd.concat([pd.DataFrame([float(mesure['profondeur_nappe'])*(-1)], columns=['piezo']) for mesure in data],ignore_index=True)
            df.set_index('date_mesure', inplace=True)
            df.to_csv(f'../data/piezo/{file}', sep=";", encoding='utf-8', )
            print(f"Fichier {file} enregistré dans data/piezo.")
        except Exception as e:
            print(f"{code_bss} : Problème durant la requête à l'API, {e}.")
            if e == "No objects to concatenate" :
                return False
            
    else : # Si le fichier existe déjà on le mentionne.
        print(f"Le fichier {file} existe déjà.")

# Test requete
url = f"http://hubeau.eaufrance.fr/api/v1/niveaux_nappes/chroniques?code_bss=03423X0056%2F100&size=20000"
get = requests.get(url)
json = get.json()
data = json['data']
data

# Collecte de toutes les chroniques de la liste établie à partir de la bdd

# Création d'une black liste des piezos sans données accessibles
piezo_black_list = []

# Liste des code_bss de la bdd
requete = """
        SELECT code_bss FROM data_piezo 
        WHERE date_debut < '2010-07-01'
        AND date_fin > '2010-07-01'
"""
liste_bss = [piezo[0] for piezo in requete_sql(requete)]

for piezo in liste_bss :
    if f_data_piezo(piezo) == False:
        # Renvoi le nom du piezo qui n'a pas de données
        piezo_black_list.append(piezo)
    
# Suppresion des piezos de la black list de la base de données
requete_del_black_list = f"""
                        DELETE FROM data_piezo
                        WHERE code_bss IN {tuple(piezo_black_list)}
                        """
requete_sql(requete_del_black_list)
print("chargement terminé")

# Mise à jour de la base de donnéees (champ fichier_piezo)
bss_piezo = pd.Series([_.replace(".csv","") for _ in os.listdir(r"./data/piezo") if _.endswith(r".csv")], name="piezo")    
requete_fichier_piezo = f"""
        UPDATE data_piezo
        SET fichier_piezo =
        (CASE WHEN code_bss IN {tuple(bss_piezo)}
        THEN 1 ELSE 0 END)
        """
requete_sql(requete_fichier_piezo)
print("Base de données à jour")
